feat: Add `mode` prop to `Answer` component
Added a `mode` prop to the `Answer` component in `Answer.tsx` to specify the mode of the chat. This prop is passed from the parent component `App.tsx`. The `mode` prop is used to conditionally render a different URL in the `navigator.share` function, based on the value of `mode`. If `mode` is equal to `research`, the URL "https://x.com/PlentyOfClarity" is used, otherwise, the URL "https://gpt-ua.click" is used. The prop `mode` is also passed down to the child components of `Answer`. Updated the `AnswerProps` interface in `Answer.tsx` to include the `mode` prop.